[TASK] Replace FormHelper unit test with a functional test
